% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytoplankton.R
\docType{data}
\name{phytoplankton}
\alias{phytoplankton}
\title{Monthly average of phytoplankton biomass between 2007 and 2021 at three stations in the Baltic Sea}
\format{
A data frame with 2,284 rows and 6 columns:
\describe{
  \item{Month_abb}{Month abbreviation}
  \item{Year}{Year}
  \item{Station}{Name of the station}
  \item{Coordinates}{Longitude/Latitude}
  \item{Taxa}{Phytoplankton taxa}
  \item{Biomass}{Phytoplankton biomass in ugC/L}
  ...
}
}
\source{
<https://sharkweb.smhi.se/hamta-data/>
}
\usage{
phytoplankton
}
\description{
Report ...
}
\keyword{datasets}
