% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard Error}
\usage{
se(x, ...)
}
\arguments{
\item{x}{a numeric vector or an R object}

\item{...}{ellipses that refers to the na.rm of the sd function}
}
\value{
a value corresponding to the standard error of the vector
}
\description{
Standard Error
}
\examples{
x <- c(2,3,2,3,4,5,1)
se(x)
data(iris)
se(iris$Sepal.Length)
}
