% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooplankton.R
\docType{data}
\name{zooplankton}
\alias{zooplankton}
\title{Monthly average of zooplankton biomass between 2007 and 2021 at three stations in the Baltic Sea}
\format{
A data frame with 2,956 rows and 7 columns:
\describe{
  \item{Month_abb}{Month abbreviation}
  \item{Year}{Year}
  \item{Station}{Name of the station}
  \item{Coordinates}{Longitude/Latitude}
  \item{Group}{Zooplankton group}
  \item{Taxa}{Zooplankton taxa}
  \item{Biomass}{Zooplankton biomass in ug/L}
  ...
}
}
\source{
<https://sharkweb.smhi.se/hamta-data/>
}
\usage{
zooplankton
}
\description{
Report ...
}
\keyword{datasets}
